New fiters available for getRegistrationList
Filter on registration 'last updated or changed'
 - after: Return registrations exclusively after a given date
 - until: Return registrations inclusively until a given date

It would seem that in certain cases a large list of registrations
can cause the ScormCloud service to timeout before it compiles the
results and returns them to the client.

As a workaround for this problem they introduced new filters that
allow the client to reduce the list based on the registration
changed date.

Filter on registration 'last updated or changed'
 - after: Return registrations exclusively after a given date
 - until: Return registrations inclusively until a given date

Tests are not included in this commit.